Fix LTC2400 readout

+++ sw/ltc2400.h
@@ -33,4 +33,4 @@ void ltc2400_init();
 
 bool ltc2400_conversion_ready();
 
-float ltc2400_get_conversion_result(bool& dmy_fault, bool& exr_fault, uint32_t& raw_value);
+double ltc2400_get_conversion_result(bool& dmy_fault, bool& exr_fault, uint32_t& raw_value);
